requirements.txt
# Requirements needed to install
pytest-playwright >= 0.3.3
python-dotenv >= 1.0.0
pytest-reporter >= 0.5.2
pytest-reporter-html1 >= 0.8.3
pytest-ordering
pytest-xdist~=3.6.1
pytest-bdd~=8.1.0
pip~=23.2.1
wheel~=0.41.2
docutils~=0.21.2
Jinja2~=3.1.6
Mako~=1.3.9
pytest~=8.3.5
MarkupSafe~=3.0.2
pyee~=12.1.1
typing_extensions~=4.13.1
execnet~=2.1.1
pluggy~=1.5.0
exceptiongroup~=1.2.2
iniconfig~=2.1.0
certifi~=2025.1.31
gherkin-official~=29.0.0
python-slugify~=8.0.4
urllib3~=2.3.0
greenlet~=3.1.1
idna~=3.10
charset-normalizer~=3.4.1
ansi2html~=1.9.2
packaging~=24.2
parse_type~=0.6.4
parse~=1.20.2
six~=1.17.0
playwright~=1.51.0
setuptools~=68.2.0